pyproject.toml
[project]
name = "pill-identifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "ultralytics>=8.4.8",
    "uvicorn>=0.40.0",
]
